Add failing contract for Game#roll
This is where I started to think about how you probably still do
need some specs even when using contracts. Contracts are dynamic
and its weird to write "example programs" for the wrong behavior
of the program.

So I wrote a couple specs for the inputs, which in theory are
what is assumed in a contract based design anyway. And then
my "Red" phase for the contract themselves is that the score
did not increase.

Getting used to the idea of writing contracts in a dynamic way
is both interesting and challenging.