Broken contract for detecting a game with spares or strikes.
This is the point in the bowling kata where I always start to feel
like I got ahead of myself and need to sit down and decompose the
problem better. It's been a long time since I've done it, so I've
not got a mental model fresh in my mind.

I took a step back and tinkered with a scoring calculator online:
https://www.sportcalculators.com/bowling-score-calculator

I think the next thing I should focus on is frame state because
you can't actually just look at "the last two balls" to determine
a bonus.

It's been a while since I've gone bowling, but the way the score
calculator in the link above works is that the total score is the
sum of the frame scores and *not* the sum of the individual
ball scores, and that seems to make a lot of sense.

Going to rethink things a bit with that in mind, and then
work that into the design.